Add id generation for work and education schemas
This will be used to create timeline links.
They should be deterministic and unique.

Strategy is to use first 5 characters of base64 encoded SHA256 hash.

For education, start time and institution is used as input.
For work, start time and company is used as input.
These _should_ uniquely identify entries and never change.