Remove sync-request dependency
Due to how amazing the NodeJS APIs are there is not easy way to make a
synchronous HTTP request without a ton of hassle. `sync-request` used to
work, but it probably did some janky shit which breaks esbuild.

This is a glorious hack. It forks a child process and calls `curl`. It
will probably break catastrophically if curl is not installed.